compute-neighbors writes bigann neighbors
