+++ et/src/compute_neighbors/mod.rs
@@ -2,10 +2,17 @@ mod cpu;
 #[cfg(feature = "wgpu")]
 mod gpu;
 
-use std::{io, num::NonZero, path::PathBuf};
+use std::{
+    fs::File,
+    io::{self, BufWriter, Write},
+    num::NonZero,
+    path::PathBuf,
+};
 
 use clap::Args;
 use vectors::VectorSimilarity;
+
+use crate::neighbor_util::TopNeighbors;
 
 #[derive(Args)]
 pub struct ComputeNeighborsArgs {
@@ -31,8 +38,9 @@ pub struct ComputeNeighborsArgs {
 
     /// Path to neighbors file to write.
     ///
-    /// The output file will contain one row for each vector in query_vectors. Within each row
-    /// there will be neighbors_len entries of Neighbor, an (i64, f64) tuple.
+    /// The output file is written in BigANN format: a <count,neighbors_len> header followed by
+    /// one row for each vector in query_vectors, each row containing neighbors_len entries of
+    /// Neighbor, an (i64, f64) tuple.
     #[arg(short, long)]
     neighbors: PathBuf,
     /// Number of neighbors for each query in the neighbors file.
@@ -52,4 +60,19 @@ pub fn compute_neighbors(args: ComputeNeighborsArgs) -> io::Result<()> {
         return gpu::run(adapter, &args);
     }
     cpu::run(&args)
+}
+
+/// Write `results` (one entry per query) to `args.neighbors` in BigANN format: a
+/// `<count,neighbors_len>` header followed by up to `neighbors_len` [`Neighbor`] entries per row.
+fn write_neighbors(args: &ComputeNeighborsArgs, results: Vec<TopNeighbors>) -> io::Result<()> {
+    let k = args.neighbors_len.get();
+    let mut writer = BufWriter::new(File::create(&args.neighbors)?);
+    writer.write_all(&(results.len() as u32).to_le_bytes())?;
+    writer.write_all(&(k as u32).to_le_bytes())?;
+    for neighbors in results.into_iter().map(TopNeighbors::into_neighbors) {
+        for n in neighbors.into_iter().take(k) {
+            writer.write_all(&<[u8; 16]>::from(n))?;
+        }
+    }
+    Ok(())
 }
